move dto into rest folder
