+++ packages/headless/src/modules/category/rest/dto/getCategory.dto.ts
